ChatRoomRepository에서 Service 분리
